requirements.txt
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
pytesseract==0.3.10
streamlit==1.21.0
